+++ offchain/eth-input-reader/src/error.rs
@@ -2,10 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0 (see LICENSE)
 
 use axum::http::uri::InvalidUri;
-use eth_state_client_lib::error::StateServerError;
+use eth_state_fold_types::ethers::providers::{
+    Http, Provider, ProviderError, RetryClient,
+};
 use snafu::Snafu;
 use std::net::AddrParseError;
 use tonic::transport::Error as TonicError;
+use url::ParseError;
 
 use crate::machine;
 
@@ -29,8 +32,20 @@ pub enum EthInputReaderError {
     #[snafu(display("connection error"))]
     ConnectError { source: TonicError },
 
-    #[snafu(display("state server error"))]
-    StateServerError { source: StateServerError },
+    #[snafu(display("provider error"))]
+    ProviderError { source: ProviderError },
+
+    #[snafu(display("parser error"))]
+    ParseError { source: ParseError },
+
+    #[snafu(display("Provider didn't return chain id"))]
+    MissingChainId,
+
+    #[snafu(display("parser error"))]
+    BlockArchiveError {
+        source:
+            eth_block_history::BlockArchiveError<Provider<RetryClient<Http>>>,
+    },
 
     #[snafu(whatever, display("{message}"))]
     Whatever {